+++ src/tossinvest_mcp/client_factory.py
@@ -6,5 +6,9 @@ from collections.abc import Callable
 from contextlib import AbstractContextManager
 
 from tossinvest import TossInvestClient
+from tossinvest_extensions import TossInvestExtensionsClient
 
 type ClientContextFactory = Callable[[], AbstractContextManager[TossInvestClient]]
+type ExtensionsClientContextFactory = Callable[
+    [], AbstractContextManager[TossInvestExtensionsClient]
+]